refactor: replace hooks global hooks component upper
