+++ client/src/components/AboutGame/components/AboutGameContent/FeatureCards/FeatureCards.component.tsx
@@ -1,5 +1,5 @@
 import { FaGamepad, FaUsers, FaTrophy } from 'react-icons/fa';
-import { useMobileDetect } from '../../../../UI/hooks/useMobileDetect.hook';
+import { useMobileDetect } from '../../../../hooks/useMobileDetect.hook';
 import { FeatureCard } from './FeatureCard.component';
 
 export const FeatureCards = () => {
