+++ client/src/components/LatestNews/components/NewsContent/NewsCards/NewsCard.component.tsx
@@ -1,5 +1,5 @@
 import { motion } from 'framer-motion';
-import { useMobileDetect } from '../../../../UI/hooks/useMobileDetect.hook';
+import { useMobileDetect } from '../../../../hooks/useMobileDetect.hook';
 
 type NewsCardProps = {
   id: string;
