+++ client/src/components/LatestNews/components/NewsContent/NewsCards/NewsCards.component.tsx
@@ -1,4 +1,4 @@
-import { useMobileDetect } from '../../../../UI/hooks/useMobileDetect.hook';
+import { useMobileDetect } from '../../../../hooks/useMobileDetect.hook';
 import { useLatestFeatures } from '../../../hooks/useLatestFeatures';
 import { LatestFeature } from '../../../api/latestFeatures.api';
 import { NewsCard } from './NewsCard.component';
