+++ client/src/components/PracticeSection/components/PracticeContent/FeaturesList/FeatureItem.component.tsx
@@ -1,6 +1,6 @@
 import { motion } from 'framer-motion';
 import { type ReactNode } from 'react';
-import { useMobileDetect } from '../../../../UI/hooks/useMobileDetect.hook';
+import { useMobileDetect } from '../../../../hooks/useMobileDetect.hook';
 
 type FeatureItemProps = {
   icon: ReactNode;
