+++ client/src/components/PracticeSection/components/PracticeContent/FeaturesList/FeaturesList.component.tsx
@@ -1,7 +1,7 @@
 import { motion } from 'framer-motion';
 import { FaGamepad, FaTrophy, FaUsers } from 'react-icons/fa';
 import { FeatureItem } from './FeatureItem.component';
-import { useMobileDetect } from '../../../../UI/hooks/useMobileDetect.hook';
+import { useMobileDetect } from '../../../../hooks/useMobileDetect.hook';
 
 export const FeaturesList = () => {
   const isMobile = useMobileDetect();
