+++ client/src/components/UI/MobileWrapper/MobileWrapper.component.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 import { motion, MotionProps } from 'framer-motion';
-import { useMobileDetect } from '../hooks/useMobileDetect.hook';
+import { useMobileDetect } from '../../hooks/useMobileDetect.hook';
 
 interface MobileWrapperProps {
   children: ReactNode;
