+++ client/src/components/UI/PhonePreview/PhonePreview.component.tsx
@@ -1,5 +1,5 @@
 import { motion } from 'framer-motion';
-import { useMobileDetect } from '../hooks/useMobileDetect.hook';
+import { useMobileDetect } from '../../hooks/useMobileDetect.hook';
 
 type PhonePreviewProps = {
   position?: 'left' | 'right';
