+++ client/src/components/UI/SectionHeader/SectionHeader.component.tsx
@@ -1,5 +1,5 @@
 import { MobileWrapper } from '../MobileWrapper/MobileWrapper.component';
-import { useMobileDetect } from '../hooks/useMobileDetect.hook';
+import { useMobileDetect } from '../../hooks/useMobileDetect.hook';
 
 type SectionHeaderProps = {
   badge?: string;
