+++ client/src/components/WhyThisGame/components/WhyThisGameContent/CodePreview/CodePreview.component.tsx
@@ -1,5 +1,5 @@
 import { motion } from 'framer-motion';
-import { useMobileDetect } from '../../../../UI/hooks/useMobileDetect.hook';
+import { useMobileDetect } from '../../../../hooks/useMobileDetect.hook';
 import { CodeEditor } from './CodeEditor/CodeEditor.component';
 import { PhonePreview } from './PhonePreview/PhonePreview.component';
