+++ client/src/components/hooks/useMobileDetect.hook.ts
