+++ client/src/components/hooks/useOnClickOutside.hook.ts
