+++ client/src/components/hooks/useSanitize.hook.ts
